CONSERTADO AS THREADS QUE NÃO ESTAVA DEIXANDO GRAVAR QUANDO AS TABELAS SÓ TINHA UM REGISTRO. CONSERTADO TAMBÉM A FINALIZAÇÃO DA COMPOSIÇÃO, FOI NECESSÁRIO MODIFICAR O CAMPO DA TABELA ITENS_PRODUTOS_INTERNOS_PAVILHAO_KIT_LOTE

+++ src/gestor/Controle/ControleListarGravarProdutosKitCompleto.java
@@ -110,7 +110,7 @@ public class ControleListarGravarProdutosKitCompleto {
                 pDigiProd.setIdProd(conecta.rs.getInt("IdProd"));
                 pDigiProd.setDescricaoProduto(conecta.rs.getString("DescricaoProd"));
                 pDigiProd.setUnidadeProd(conecta.rs.getString("UnidadeProd"));
-                pDigiProd.setQuantidadeProd(conecta.rs.getFloat("QuantProd"));
+                pDigiProd.setQuantidadeProd(conecta.rs.getInt("QuantProd"));
                 listaInternosPavilhaoSelecionados.add(pDigiProd);
                 qtdProd = qtdProd + 1;
             }
