+++ src/gestor/Controle/ControleProdutosKitLote.java
@@ -109,7 +109,7 @@ public class ControleProdutosKitLote {
                 pDigiProd.setIdProd(conecta.rs.getInt("IdProd"));
                 pDigiProd.setDescricaoProduto(conecta.rs.getString("DescricaoProd"));
                 pDigiProd.setUnidadeProd(conecta.rs.getString("UnidadeProd"));
-                pDigiProd.setQuantidadeProd(conecta.rs.getFloat("QuantProd"));
+                pDigiProd.setQuantidadeProd(conecta.rs.getInt("QuantProd"));
                 listaInternosPavilhaoSelecionados.add(pDigiProd);
                 qtdProd = qtdProd + 1;
             }
